+++ src/domain/ListEventStreamRecord.ts
